"delete: add swDest in workbox config"
