+++ src/pages/support/Support.tsx
@@ -1,13 +1,42 @@
-import { Block, Page } from 'framework7-react';
-import { NotifNav } from './components';
-import SupportComponent from './components/SupportComponent';
+import { Block, Page, Button } from 'framework7-react';
+import { SupportNav } from './components';
+import TelegramIcon from '@/assets/image/telegram.svg';
+import MeiQiaIcon from '@/assets/image/customer_service1.svg';
+
+type SupportChannel = {
+  label: string;
+  handle: string;
+  icon: string;
+};
+
+const CHANNELS: SupportChannel[] = [
+  { label: 'telegram', handle: '@telegram', icon: TelegramIcon },
+  { label: 'meiQia', handle: 'Chinese', icon: MeiQiaIcon },
+];
+
+const SupportCard = ({ label, handle, icon }: SupportChannel) => (
+  <>
+    <p className="text-gradient font-bold">{label}</p>
+    <div className="flex items-center justify-between gap-6 rounded-lg bg-white px-5 py-5">
+      <div className="flex items-center space-x-2">
+        <img src={icon} alt={label} className="h-7 w-7" />
+        <p className="text-base">{handle}</p>
+      </div>
+      <Button className="h-9 rounded-lg bg-primary-gradient text-white">Chat Now</Button>
+    </div>
+  </>
+);
 
 const SupportPage = () => {
   return (
     <Page name="Support">
-      <NotifNav title="24H Customer Service" />
+      <SupportNav title="24H Customer Service" />
       <Block>
-        <SupportComponent />
+        <div className="space-y-1 pb-20">
+          {CHANNELS.map((c) => (
+            <SupportCard key={c.label} {...c} />
+          ))}
+        </div>
       </Block>
     </Page>
   );
