+++ src/pages/support/components/index.ts
@@ -1,2 +1,1 @@
 export * from './SupportNavbar';
-export * from './SupportComponent';
